Move titlebar to be with extensions
