+++ src/browser/Curriculum.tsx
@@ -16,6 +16,7 @@ import * as cai from "../cai/caiState"
 import * as caiThunks from "../cai/caiThunks"
 import { Language } from "common"
 import { selectExtensionIcon32, selectExtensionName } from "../extensions/extensionState"
+import { TitleBar } from "../extensions/TitleBar"
 
 const SECTION_URL_CHARACTER = ":"
 
@@ -196,31 +197,6 @@ const CurriculumSearchResults = () => {
         : null
 }
 
-const TitleBar = ({ title, closeButtonTitle, children }: { title: string, closeButtonTitle: string, children?: React.ReactNode }) => {
-    const dispatch = useDispatch()
-
-    return (
-        <div className="flex items-center p-2 text-base bg-white text-black dark:bg-gray-900 dark:text-white">
-            <div className="ltr:pl-2 ltr:pr-4 rtl:pl-4 rtl:pr-3 font-semibold truncate">
-                <h2>{title.toLocaleUpperCase()}</h2>
-            </div>
-            <div>
-                <button
-                    className="flex justify-end w-7 h-4 p-0.5 rounded-full cursor-pointer bg-black dark:bg-gray-700"
-                    onClick={() => dispatch(layout.setEast({ open: false }))}
-                    title={closeButtonTitle}
-                    aria-label={closeButtonTitle}
-                >
-                    <div className="w-3 h-3 bg-white rounded-full">&nbsp;</div>
-                </button>
-            </div>
-            <div className="flex items-center ltr:ml-auto rtl:mr-auto">
-                {children}
-            </div>
-        </div>
-    )
-}
-
 export const CurriculumTitleBar = () => {
     const { t } = useTranslation()
     const dispatch = useDispatch()
@@ -257,22 +233,6 @@ export const CurriculumTitleBar = () => {
                     <img src={extensionIcon32} alt="" className="w-5 h-5" />
                 </button>
             )}
-        </TitleBar>
-    )
-}
-
-export const ExtensionsTitleBar = () => {
-    const { t } = useTranslation()
-    const dispatch = useDispatch()
-
-    return (
-        <TitleBar title={t("extensions")} closeButtonTitle={t("extension.close")}>
-            <button
-                className="inline-flex items-center p-1 text-xs rounded bg-gray-100 dark:bg-gray-800 border border-gray-300 dark:border-gray-600 shadow-inner hover:bg-gray-200 dark:hover:bg-gray-700 ml-2"
-                title={t("curriculum.title")}
-                onClick={() => { dispatch(appState.setEastContent("curriculum")) }}>
-                {t("curriculum.title").toLocaleUpperCase()}
-            </button>
         </TitleBar>
     )
 }
